+++ src/linear_algebra/Vector.hpp
@@ -27,7 +27,7 @@ void copy_from(std::vector<double>& destination, const std::vector<double>& sour
 double norm_1(const std::vector<double>& x);
 double norm_1(const SparseVector& x);
 double norm_1(const std::vector<SparseVector>& m);
-double norm_1(const std::function<double(int i)>& f, size_t size);
+double norm_1(const std::function<double(size_t i)>& f, size_t size);
 
 double norm_2_squared(const std::vector<double>& x);
 double norm_2_squared(const SparseVector& x);
@@ -65,7 +65,7 @@ double norm(const T& x, Norm norm) {
     }
 }
 
-double norm(const std::function<double(int i)>& f, size_t size, Norm norm);
+double norm(const std::function<double(size_t i)>& f, size_t size, Norm norm);
 
 template <typename T>
 void print_vector(std::ostream &stream, const std::vector<T>& x, const char end='\n', size_t start = 0, size_t length = std::numeric_limits<size_t>::max()) {
